+++ wrapperfunction/admin/service/crawl_service.py
@@ -80,7 +80,7 @@ def orchestrator_function(
         json_data = json.dumps(site_data, ensure_ascii=False)
         blob_name = f"item_{process_text_name(url.link)}.json"
         append_blob(
-            folder_name=config.SUBFOLDER_NAME,
+            folder_name= base_url if settings.mediaCrawling else config.SUBFOLDER_NAME,
             blob_name=blob_name,
             blob=json_data,
             container_name=settings.containerName,
