+++ wrapperfunction/core/config.py
@@ -74,6 +74,8 @@ SEARCH_API_VERSION=os.getenv("SEARCH_API_VERSION")
 COSMOS_MEDIA_KNOWLEDGE_TABLE=os.getenv("COSMOS_MEDIA_KNOWLEDGE_TABLE")
 X_KEY=os.getenv("X_KEY")
 X_TABLE=os.getenv("X_TABLE")
+MOST_INDEXED_URLS_TABLE=os.getenv("MOST_INDEXED_URLS_TABLE")
+MOST_USED_KEYWORDS_TABLE=os.getenv("MOST_USED_KEYWORDS_TABLE")
 
 
 def load_entity_settings():
